# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dusty Jewelry Store API",
        "description": "API documentation for Dusty Jewelry Store management system",
        "contact": {
            "email": "support@dusty.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get all expenses",
                "operationId": "8bcf0194356bbaeef047c66188f65944",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of expenses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Create a new expense",
                "operationId": "3ea1b5dbe09af5276ecb56173ce789bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_id",
                                    "expense_type_id",
                                    "amount",
                                    "expense_date"
                                ],
                                "properties": {
                                    "branch_id": {
                                        "type": "integer"
                                    },
                                    "expense_type_id": {
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "expense_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "notes": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/expenses/{id}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get a specific expense",
                "operationId": "64b051c55aa8ec7bc35dbdbeec5a1358",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense details"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/comprehensive": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get comprehensive report",
                "operationId": "9731f7481ade00795f449c46a790fe5f",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "branch",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comprehensive report data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reports/dashboard-stats": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get dashboard statistics",
                "operationId": "a06453561be5191bbfd11bb54553062a",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Get all sales",
                "operationId": "248121b133c730c6860e94383dfe3e12",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of sales"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Create a new sale",
                "operationId": "42cd0cb8dcc906da8e52634de671f7bf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "branch_id",
                                    "employee_id",
                                    "products",
                                    "payment_method"
                                ],
                                "properties": {
                                    "branch_id": {
                                        "type": "integer"
                                    },
                                    "employee_id": {
                                        "type": "integer"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "network",
                                            "mixed"
                                        ]
                                    },
                                    "cash_amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "network_amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "category_id": {
                                                    "type": "integer"
                                                },
                                                "caliber_id": {
                                                    "type": "integer"
                                                },
                                                "weight": {
                                                    "type": "number"
                                                },
                                                "amount": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sale created successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales/{id}": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Get a specific sale",
                "operationId": "b8c1920044f4a939f5bd289d5693854a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sale details"
                    },
                    "404": {
                        "description": "Sale not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales/{id}/return": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Return a sale",
                "operationId": "c444a1f8a4fcdb9c66da1df08faf6d54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sale returned successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sales/search/invoice": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Search sales by invoice number",
                "operationId": "93273f13a3bc08d19e125882219d6ae6",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}